Enemy Initialization/basic functionality
Further progress on setting up the enemy stats/types interactions, as
well as the levels behavior.  There is now a level area(which is
independant of the camera) that each type is notified of, there is a
"strong" enemy type that has more health but less speed, and most
importantly enemies are being created and are drawn on the screen.